move objs and group funcs to base page
